+++ 4/test/TetsPhysicalMemory.cpp
@@ -4,7 +4,7 @@
 #define CATCH_CONFIG_MAIN  // This tells Catch to provide a main() - only do this in one cpp file
 
 #include <catch2/catch.hpp>
-#include "../src/PhysicalMemory.h"
+#include "../src/file-system/PhysicalMemory.h"
 
 TEST_CASE("test physical memory CRD") {
     PhysicalMemory physicalMemory(std::string(100, ' '));
